requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
black==22.6.0
click==8.1.3
Deprecated==1.2.13
Django==4.0.6
django-cors==0.1
django-cors-headers==3.13.0
django-countries==7.3.2
django-environ==0.9.0
django-ninja==0.19.0
django-redis==5.2.0
djangorestframework==3.13.1
dnspython==2.2.1
email-validator==1.2.1
environ==1.0
filelock==3.7.1
idna==3.3
mypy-extensions==0.4.3
packaging==21.3
password-validator==1.0
pathspec==0.9.0
platformdirs==2.5.2
pydantic==1.9.1
PyJWT==2.4.0
pyparsing==3.0.9
python-usernames==0.3.1
pytz==2022.1
redis==4.3.4
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.3.0
validate-email==1.3
wrapt==1.14.1
